+++ CHANGELOG.md
@@ -4,6 +4,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [unreleased](https://github.com/uzh/z-pool-tool/tree/HEAD)
 
+### Added
+
+- Option to customize the text on the unsubscribe page using the new texts "Unsubscribe 
+  Title" and "Unsubscribe Text"
+
 ## [0.13.9](https://github.com/uzh/z-pool-tool/tree/0.13.9)
 
 ### Changed